--- src/algorithm/z-test.jl
@@ -1,8 +1,0 @@
-# algorithm/z-test.jl
-
-using Test: @test
-
-@test ncg(:test)
-@test ogm_ls(:test)
-@test pogm_restart(:test)
-@test poweriter(:test)
