+++ src/fbp/ellipse_sino.jl
@@ -6,7 +6,6 @@ ellipse_sino.jl
 
 export ellipse_sino
 
-using Plots: plot
 #using MIRT: sino_geom, MIRT_sino_geom, downsample2
 
 
@@ -111,88 +110,4 @@ function ellipse_sino(rg::AbstractArray{<:Real}, ϕg::AbstractArray{<:Real},
 	end
 
 	return sino
-end
-
-
-"""
-`ellipse_sino()`
-
-show doc strings
-"""
-function ellipse_sino()
-	@doc ellipse_sino
-end
-
-
-"""
-`ellipse_sino_test()`
-
-internal test routine: standard sampling
-"""
-function ellipse_sino_test()
-	ell = [ 40 70 50 150 20 10 ]
-	sg = sino_geom(:ge1, down=8)
-	ellipse_sino(sg, ell; xscale=-1, yscale=-1) # test scale
-	true
-end
-
-
-"""
-`ellipse_sino_show()`
-show examples
-
-To see more, use `MIRT.ellipse_sino_show(ip=2)`
-"""
-function ellipse_sino_show( ;
-		down::Int = 4,
-		ell::Matrix = [ 40 60 50 150 20 10 ],
-		orbit::Real = 360,
-		na::Int = 400,
-		oversample::Int = 2,
-		ip::Int = 1,
-	)
-
-	ig = image_geom(nx=512, ny=504, dx=1, dy=1, mask=:circ)
-	ig = ig.down(down)
-	xtrue = ellipse_im(ig, ell ; oversample=2)
-	ig = image_geom(nx=ig.nx, ny=ig.ny, dx=ig.dx, dy=ig.dy, mask = xtrue .> 0)
-
-	geoms = (
-		sino_geom(:par, nb = 444, na=na, down=down, d = 1, orbit=orbit,
-			offset = 0.25),
-		sino_geom(:fan, nb = 888, na=na, d = 1.0, orbit=orbit,
-			offset = 0.75, dsd = 949, dod = 408, down=down),
-		sino_geom(:fan, nb = 888, na=na, d = 1.0, orbit=orbit,
-			offset = 0.75, dsd = 949, dod = 408, down=down,
-			dfs = Inf, source_offset = 0.7), # flat fan
-		sino_geom(:moj, nb = 630, na=na, down=down, d = 1.0, orbit=orbit,
-			offset = 0.25),
-	)
-
-	ngeom = length(geoms)
-	pl = Array{Plot}(undef, ngeom, 2)
-
-	for ii=1:ngeom
-		sg = geoms[ii]
-		sino = ellipse_sino(sg, ell; oversample=oversample)
-		dfs = sg.how === :fan ? " dfs=$(sg.dfs)" : ""
-		pl[ii,1] = jim(sino, title="$(sg.how)$dfs")
-		pl[ii,2] = sg.plot(ig=ig)
-	end
-#	plot(pl..., layout=(2,ngeom)) # too small
-	plot(pl[:,ip]...)
-end
-
-
-"""
-`ellipse_sino(:test)`
-self test
-"""
-function ellipse_sino(test::Symbol)
-	test === :show && return ellipse_sino_show()
-	test != :test && throw(ArgumentError("test $test"))
-	ellipse_sino() # doc
-	ellipse_sino_test()
-	ellipse_sino(:show)
-	true
 end
